src/db/connection/database.py
from typing import AsyncGenerator

from sqlalchemy.ext.asyncio import create_async_engine
from sqlalchemy.orm import sessionmaker
from sqlmodel.ext.asyncio.session import AsyncSession

from src.config import config

DATABASE_URL = str(config.database_url)

engine = create_async_engine(DATABASE_URL, echo=False, future=True)


async_session_maker = sessionmaker(
    engine,
    class_=AsyncSession,
    expire_on_commit=False,
    autocommit=False,
    autoflush=False,
)


async def get_session() -> AsyncGenerator[AsyncSession, None]:

    async with async_session_maker() as session:
        yield session
