requirements.txt
fastapi
uvicorn[standard]
ruff
alembic
sqlmodel
asyncpg
pwdlib[argon2]
redis
PyJWT
python-multipart
pydantic-settings